+++ package.json
@@ -44,6 +44,8 @@
     "eslint-plugin-react": "^7.37.4",
     "globals": "^15.14.0",
     "postcss": "^8.5.1",
+    "prettier": "^3.4.2",
+    "prettier-plugin-tailwindcss": "^0.6.11",
     "tailwindcss": "^3.4.17"
   },
   "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
