+++ src/App.js
@@ -5,9 +5,7 @@ import Router from "./router/router";
 function App() {
   return (
     <BrowserRouter>
-      {/* <Header />   */}
       <Router />
-      {/* <Footer /> */}
     </BrowserRouter>
   );
 }
