+++ src/components/account/ChangePW.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function ChangePW() {
+  return <div>ChangePW</div>;
+}
