+++ src/components/account/DeleteID.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function DeleteID() {
+  return <div>DeleteID</div>;
+}
