+++ src/components/account/Login.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Login() {
+  return <div>Login</div>;
+}
