+++ src/components/account/SignUp.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function SignUp() {
+  return <div>SignUp</div>;
+}
