+++ src/components/common/FilteredList.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function FilteredList() {
+  return <div>FilteredList</div>;
+}
