+++ src/components/common/OneBtnModal.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function OneBtnModal() {
+  return <div>OneBtnModal</div>;
+}
