+++ src/components/common/TwoBtnModal.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function TwoBtnModal() {
+  return <div>TwoBtnModal</div>;
+}
