+++ src/components/layout/Header.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Header() {
+  return <div>Header</div>;
+}
