+++ src/components/matching/AgreePlace.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function AgreePlace() {
+  return <div>AgreePlace</div>;
+}
