+++ src/components/matching/ChoicePlace.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function ChoicePlace() {
+  return <div>ChoicePlace</div>;
+}
