+++ src/components/matching/MatchingComplete.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function MatchingComplete() {
+  return <div>MatchingComplete</div>;
+}
