+++ src/components/rests/RestInfo.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function RestInfo() {
+  return <div>RestInfo</div>;
+}
