+++ src/components/rests/RestList.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function rests() {
+  return <div>Rests</div>;
+}
