+++ src/components/rests/RestReviews.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function RestReviews() {
+  return <div>RestReviews</div>;
+}
