+++ src/components/rests/RestView.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function rest() {
+  return <div>Rest</div>;
+}
