+++ src/components/rests/WriteReview.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function WriteReview() {
+  return <div>WriteReview</div>;
+}
