+++ src/views/Account.js
@@ -1,0 +1,10 @@
+import React from "react";
+import Login from "../components/account/Login";
+
+export default function Account() {
+  return (
+    <div>
+      <Login />
+    </div>
+  );
+}
