+++ src/views/MyPage.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function MyPage() {
+  return <div>MyPage</div>;
+}
