+++ src/views/OpenChat.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function OpenChat() {
+  return <div>OpenChat</div>;
+}
